feat: allow creation of multi-key indices
`bson.M` cannot be used to create multi-key indices (they don't guarantee the order of the keys, which can make a difference on mongo indices). The mongodb driver does not allow using `bson.M` for that, it features a protection against that (https://jira.mongodb.org/browse/GODRIVER-1810).

This commit adds a `CreateMultiKeyIndex` func, which gets `bson.D`, which is a set of ordered key-value pairs. This is allowed and works as expected.